requirements.txt
langchain-groq
langchain
langchain-groq
python-dotenv
langchain_community
PyMuPDF
python-dotenv
pypdf